requirements.txt
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
DateTime==4.3
decorator==4.4.0
Flask==1.0.2
gunicorn==19.9.0
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-core==4.4.0
MarkupSafe==1.1.1
nbformat==4.4.0
numpy==1.16.3
pandas==0.24.2
plotly==3.9.0
psycopg2-binary==2.8.2
py4j==0.10.7
pyrsistent==0.15.2
pyspark==2.4.0
python-dateutil==2.8.0
python-twitch-client==0.6.0
pytz==2019.1
requests==2.21.0
retrying==1.3.3
six==1.12.0
SQLAlchemy==1.3.3
traitlets==4.3.2
urllib3==1.24.3
Werkzeug==0.15.3
zope.interface==4.6.0
